menu highlight
Former-commit-id: 103a22762d49fc56cc09ba8f3ba97b82130f946f [formerly 2fa3a495b6e0d3af3650e7545e6ce04ee52cbdac] [formerly c8e3f6cbc7ad30ae9c7875a7320f1643d8315981 [formerly fe489fcde3efdf624ab9f4a73a55d2983151fce9]] [formerly 5bab1b65dd73fd12e5a5bbb2b38d79ee21a4725f [formerly 71b875634459c1834f45b84f54a95b49944aa0f9] [formerly 7fae9f028a325bdee09c6296776d6f38aa6e98d8 [formerly dbc945330a2a64b243b8f47b2c6e7326c0971403]]]
Former-commit-id: 0078470cf9a7fbf2542bf22110dd413890f3186f [formerly 8b400c28eca9e5b4dd557538c04cfad8a2126e13] [formerly 0efbde14dc650a018261cce000d7a0c0c20b7e8b [formerly 3b0e6f1a32274ae6eea04408edaf72dc99cf334e]]
Former-commit-id: d6aa4c41820d1c4ecebf1e1bed4375cd557d9eb3 [formerly 7e8638a598eaa5c5d61e678702ca6d807dbfe559]
Former-commit-id: 7007a44f9b5246b8e38ebf76f01bc42fcb8288da

+++ Assets/Scripts/Decal/DecalScript.cs
@@ -6,7 +6,7 @@ public class DecalScript : MonoBehaviour
 {
     public float TotalTimeAlive = 2F;
     public float CurrentTime = 0F;
-    private float AliveCounter = 0F;
+    private int AliveCounter = 0;
     // Start is called before the first frame update
     void Start()
     {
@@ -26,7 +26,7 @@ public class DecalScript : MonoBehaviour
 
     void Reposition()
     {
-        AliveCounter = 0F;
+        AliveCounter = 0;
         Ray ray = new Ray(transform.position, transform.forward);
         RaycastHit hit;
         if (Physics.Raycast(ray, out hit, 3F))
@@ -41,12 +41,12 @@ public class DecalScript : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        if(CurrentTime>TotalTimeAlive)
+        if (CurrentTime > TotalTimeAlive)
         {
             AliveCounter++;
         }
 
-        if(AliveCounter>10)
+        if (AliveCounter > 10)
         {
             PrefabPool.Despawn(gameObject);
         }
